chore(lazygit): remove version field
After reviewing the GitHub repository for kdheepak/lazygit.nvim, it
appears that the plugin does not have any official releases or tags.
This means that specifying a version like version = "v1.2.3" or even
version = "*" in plugin configuration won’t have any effect, as there
are no versioned releases to reference.